+++ lib/bra_document_validation.rb
@@ -5,6 +5,7 @@ require 'bra_documents'
 require 'bra_document_validation/version'
 require 'bra_document_validation/cnpj_validator'
 require 'bra_document_validation/cpf_validator'
+require 'bra_document_validation/cpf_or_cnpj_validator'
 
 module BraDocumentValidation
 end