+++ lib/bra_document_validation/cpf_or_cnpj_validator.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module BraDocumentValidation
-  class CPFValidator < ActiveModel::EachValidator
+  class CPFOrCNPJValidator < ActiveModel::EachValidator
     def validate_each(record, attribute, value)
       return record.errors.add(attribute, error_message(:invalid_format)) unless valid_format?(value)
 
@@ -11,11 +11,13 @@ module BraDocumentValidation
     private
 
     def valid_format?(document)
-      BraDocuments::Matcher.match?(document.to_s, kind: :cpf, mode: document_format)
+      BraDocuments::Matcher.match?(document.to_s, kind: :cpf, mode: document_format) ||
+        BraDocuments::Matcher.match?(document.to_s, kind: :cnpj, mode: document_format)
     end
 
     def valid_verification_digit?(document)
-      BraDocuments::CPFGenerator.valid_verification_digit?(document: document.to_s)
+      BraDocuments::CPFGenerator.valid_verification_digit?(document: document.to_s) ||
+        BraDocuments::CNPJGenerator.valid_verification_digit?(document: document.to_s)
     end
 
     def document_format