+++ CHANGELOG.md
@@ -1,18 +1,27 @@
 
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Removed
+
 - Drop Support for Ruby 2.6 or less
+- Refactory CPF and CNPJ Validators to use new bra_documents's methods
+- Add new validator to check if an attribute is a CPF or a CNPJ
 
 ## [1.0.1] - 2020-11-08
+
 ### Fixed
+
 - Fix Gemfile lock
 
 ## [1.0.0] - 2020-11-08
+
 ### Added
+
 - CPF and CNPJ Validators